Move http module to waltz library
fd_waltz contains all networking code.
Aligns the http module accordingly.
