+++ src/disco/gui/fd_gui.h
@@ -4,7 +4,7 @@
 #include "../fd_disco_base.h"
 
 #include "../pack/fd_microblock.h"
-#include "../../ballet/http/fd_http_server.h"
+#include "../../waltz/http/fd_http_server.h"
 #include "../../flamenco/types/fd_types.h"
 #include "../../flamenco/leaders/fd_leaders.h"
