+++ src/disco/quic/Local.mk
@@ -3,7 +3,7 @@ $(call add-objs,fd_tpu_reasm,fd_disco)
 $(call make-unit-test,test_tpu_reasm,test_tpu_reasm,fd_disco fd_tango fd_ballet fd_util)
 $(call run-unit-test,test_tpu_reasm)
 ifdef FD_HAS_DOUBLE
-$(call make-unit-test,test_quic_metrics,test_quic_metrics,fd_disco fd_tango fd_ballet fd_util)
+$(call make-unit-test,test_quic_metrics,test_quic_metrics,fd_disco fd_waltz fd_tango fd_ballet fd_util)
 $(call run-unit-test,test_quic_metrics)
 $(OBJDIR)/obj/disco/quic/test_quic_metrics.o: src/disco/quic/test_quic_metrics.txt
 endif
