+++ src/flamenco/snapshot/Local.mk
@@ -5,10 +5,10 @@ ifdef FD_HAS_INT128
 ifdef FD_HAS_HOSTED
 $(call add-hdrs,fd_snapshot_http.h)
 $(call add-objs,fd_snapshot_http,fd_flamenco)
-$(call make-unit-test,test_snapshot_http,test_snapshot_http,fd_flamenco fd_disco fd_ballet fd_util)
+$(call make-unit-test,test_snapshot_http,test_snapshot_http,fd_flamenco fd_disco fd_waltz fd_ballet fd_util)
 $(call run-unit-test,test_snapshot_http)
 ifdef FD_HAS_THREADS
-$(call make-fuzz-test,fuzz_snapshot_http,fuzz_snapshot_http,fd_flamenco fd_disco fd_ballet fd_util)
+$(call make-fuzz-test,fuzz_snapshot_http,fuzz_snapshot_http,fd_flamenco fd_disco fd_waltz fd_ballet fd_util)
 endif
 endif
 
@@ -32,6 +32,6 @@ $(call add-objs,fd_snapshot_loader,fd_flamenco)
 $(call add-hdrs,fd_snapshot_create.h)
 $(call add-objs,fd_snapshot_create,fd_flamenco)
 
-$(call make-bin,fd_snapshot,fd_snapshot_main,fd_flamenco fd_disco fd_funk fd_ballet fd_util,$(SECP256K1_LIBS))
+$(call make-bin,fd_snapshot,fd_snapshot_main,fd_flamenco fd_disco fd_funk fd_waltz fd_ballet fd_util,$(SECP256K1_LIBS))
 endif
 endif
