+++ src/waltz/http/test_http_server.c
@@ -1,5 +1,6 @@
 #include "fd_http_server.h"
 #include "fd_http_server_private.h"
+#include "../../util/fd_util.h"
 
 void
 test_oring( void ) {
